backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mem0ai
openai
python-dotenv
qdrant-client